requirements.txt
boto==2.38.0
colorama==0.3.5
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-haystack==2.4.1
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==3.3.1
gunicorn==19.4.1
httpie==0.9.2
psycopg2==2.6.1
Pygments==2.0.2
pytz==2015.7
requests==2.8.1
static3==0.6.1
wheel==0.24.0
